TOT-99 : [FIX] 지역 제거시 제거된 장소도 제거되도록 수정
