+++ frontend/src/@types/api/placeList.ts
@@ -6,4 +6,5 @@ export interface PlaceListTypes {
   latitude: number;
   longitude: number;
   contentTypeId: number;
+  regionId?: string;
 }
