+++ frontend/src/components/RegionCategoryList/index.ts
@@ -1,0 +1,3 @@
+import RegionCategoryList from './RegionCategoryList';
+
+export default RegionCategoryList;
